### 2020-12-04 #2884 @kyleniemeyer
Hello @jbytecode, thanks for your interest in JOSS. We will use this pre-review issue to identify an editor and reviewers. While doing that, could you fix those issues related to DOIs in the paper? I think you just need to remove `https://doi.org/` at the beginning and just include the DOI itself in the `.bib` file.
