### 2020-12-05 c21e624
remove invalid doi enteries
