+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages     = {726 - 740},
   year      = {1987},
   isbn      = {978-0-08-051581-6},
-  doi       = {https://doi.org/10.1016/B978-0-08-051581-6.50070-2},
+  doi       = {10.1016/B978-0-08-051581-6.50070-2},
   url       = {http://www.sciencedirect.com/science/article/pii/B9780080515816500702},
   author    = {Martin A. Fischler and Robert C. Bolles},
   abstract  = {A new paradigm, Random Sample Consensus (RANSAC), for fitting a model to experimental data is introduced, RANSAC is capable of interpreting/ smoothing data containing a significant percentage of gross errors, and is thus ideally suited for applications in automated image analysis where interpretation is based on the data provided by error-prone feature detectors. A major portion of this paper describes the application of RANSAC to the Location Determination Problem (LDP): Given an image depicting a set of landmarks with known locations, determine that point in space from which the image was obtained. In response to a RANSAC requirement, new results are derived on the minimum number of landmarks needed to obtain a solution, and algorithms are presented for computing these minimum-landmark solutions in closed form. These results provide the basis for an automatic system that can solve the LDP under difficult viewing and analysis conditions. Implementation details and computational examples are also presented.}
@@ -26,7 +26,7 @@
 
 @article{ks89,
   issn      = {0006341X, 15410420},
-  doi       = {https://doi.org/10.2307/2531498},
+  doi       = {10.2307/2531498},
   url       = {http://www.jstor.org/stable/2531498},
   author    = {Farid Kianifard and William H. Swallow},
   journal   = {Biometrics},
@@ -47,7 +47,7 @@
   number   = {3},
   pages    = {761-771},
   keywords = {leverage points, mahalanobis distance, masking, minimum volume ellipsoid, robust distance, swamping},
-  doi      = {https://doi.org/10.1111/j.2517-6161.1992.tb01449.x},
+  doi      = {10.1111/j.2517-6161.1992.tb01449.x},
   url      = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1992.tb01449.x},
   eprint   = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.2517-6161.1992.tb01449.x},
   year     = {1992}
@@ -86,15 +86,7 @@
   pages     = {1329-1339},
   year      = {1994},
   publisher = {Taylor & Francis},
-  doi       = {10.1080/01621459.1994.10476872},
-  url       = { 
-        https://doi.org/10.1080/01621459.1994.10476872
-    
-},
-  eprint    = { 
-        https://doi.org/10.1080/01621459.1994.10476872
-    
-}
+  doi       = {10.1080/01621459.1994.10476872}
 }
 
 
@@ -106,7 +98,7 @@
   number   = {2},
   pages    = {393-396},
   keywords = {mahalanobis distance, masking, power, robust statistics, swamping},
-  doi      = {https://doi.org/10.1111/j.2517-6161.1994.tb01988.x},
+  doi      = {10.1111/j.2517-6161.1994.tb01988.x},
   url      = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1994.tb01988.x},
   eprint   = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.2517-6161.1994.tb01988.x},
[…]
