+++ paper/paper.bib
@@ -15,7 +15,7 @@
 
 @article{julia,
   title     = {Julia: A fresh approach to numerical computing},
-  doi       = {10.1137/141000671}
+  doi       = {10.1137/141000671},
   author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal   = {SIAM review},
   volume    = {59},
