### 2020-12-08 #2892 @whedon
[REVIEW]: LinRegOutliers: A Julia package for detecting outliers in linear regression
**Submitting author:** @jbytecode (<a href="http://orcid.org/0000-0002-9402-1982">Mehmet Hakan Satman</a>)
**Repository:** <a href="https://github.com/jbytecode/LinRegOutliers" target ="_blank">https://github.com/jbytecode/LinRegOutliers</a>
**Version:** v0.8.5
