### 2020-12-31 #2892 @salleuska
## Comments on the software repository

I have some concerns regarding documentation of functions and examples, but I think these can be easily addressed by the author.  My main concern is that the documentation should be expanded. Although each of the methods seems well referenced, few sentences describing what is the base idea of each algorithm would be helpful in my opinion. This also should help to clarify why some functions return different outputs (`Array` vs `Dictionary`) and what each element of the `Dictionary` is.

I'll open a few more detailed issues in the repository to reference here, and make a separate comment related to Software Paper checklist.

I admit that I have less familiarity with `Julia` packages than `R`, and what I am suggesting is what is typically done in `R`. If @mikldk or @rMassimiliano have some different view, please let me know.
