### 2020-12-31 jbytecode/LinRegOutliers#40 @salleuska
Documentation - 1. Functions
I think that documentation related to the algorithms provided in the package needs to be improved. Although each method has its clear reference provided in the [reference manual](https://jbytecode.github.io/LinRegOutliers/docs/build/algorithms/), I think it would be nice to have a couple of sentences providing an idea of what the algorithm does.

In addition, outputs from different algorithms are not consistent between each other. This not a problem per se, but I think that there should be a description of what the different output elements are. For example `ks89` and `smr98` seems to return an `Array` of indexes for the outliers, while `lms`returns a `Dictionary` with 6 entries. These entries are not obvious to interpret, and I think should be documented.

https://github.com/openjournals/joss-reviews/issues/2892
