### 2020-12-31 jbytecode/LinRegOutliers#41 @salleuska
Documentation - 2. Examples
As for the documentation, I think that examples could clearer if expanded a bit. I like the fact that algorithms calls follow the same structure, the function usage is pretty self-explanatory.
However, I would suggest expanding the [example page](https://github.com/jbytecode/LinRegOutliers/blob/master/examples.md) to better reflect how the functions of the package help to perform data analysis. For example, providing few sentences describing the data, objective of the analysis and commenting the results. The example could also clarify that algorithms can provide different outputs, and that they all provide a list of outliers named `outliers` directly or contained in the `Dictionary`.

https://github.com/openjournals/joss-reviews/issues/2892
