### 2020-12-31 jbytecode/LinRegOutliers#42 @salleuska
Consistency about objects returned from functions
I am not sure this is strictly related with the review, but I have noticed that `satman2013` returns a `Dictionary` with 1 entry named `outliers`, while functions `ks89` and `smr98` returns an `Array`. Aren't all these functions returning indexes of outliers? If so, I think it would be nice that the objects returned are consistent across functions (either a `Dictionary` with one entry or an `Array`).

https://github.com/openjournals/joss-reviews/issues/2892
