### 2020-12-31 #2892 @salleuska
## Software paper

I think that the paper does not require major changes in the structure or language. I have just a few minor comments

- **Summary**. ``Our package covers a significant portion of the literature on *fitting with outliers*'' -> Our package [..] on *(linear regression | model fitting)* with outliers
- **Statement of need**. I wonder if the methods can be listed or tabulated according to the 4 categories mentioned in the previous section (diagnostics, direct methods, robust methods, multivariate). This should help in navigating between the many methods.
- **Installation and basic usage**. It is explained that some methods may return additional information, but there is no documentation or reference to what this information is. I think it would be better this information to be documented in the package as suggested in my previous issues. If not should be made clear that one has to look into the paper related to the method.
