### 2020-12-31 0081a22
consistent return type of Dict for all exported methods
