### 2020-12-31 jbytecode/LinRegOutliers#42 @jbytecode
Dear reviewer,

By the last [commit](https://github.com/jbytecode/LinRegOutliers/commit/0081a2212c77bbc1f58e02e3ac8bb56e5d36d3bb), all of the exported methods consistently return a `Dict`.

Thank you.
