### 2020-12-31 9edf826
add explanations in examples.md
