+++ examples.md
@@ -1,51 +1,76 @@
 # Some examples
 
+
+## A brief introduction
+
+Suppose the linear regression model is 
+
+*y = Xβ + ε*
+
+where *y* is the vector of dependent variable, *X* is the *n ˟ p* matrix of design, *ε* is i.i.d error term with zero mean, *n* is the number of observations, and *p* is the number of regression parameters. 
+
+When a single regressor exists in the model, it can be basically written as 
+
+*y = β₀ + β₁ x +  ε*
+
+where *β₀* and *β₁* are unknown intercept and slope parameters. In `R` and `Julia` we can represent this model in a similar form. Specifically, in Julia, the simple model can be expressed using the `@formula` macro as
+
+```julia
+@formula(y ~ x)
+```
+
+where ```~``` operator seperates the dependent and independent variables. When the model includes more than one regressors, the model can similarly be expressed as
+
+```julia
+@formula(y ~ x1 + x2 + x3)
+```
+
+```LinRegOutliers``` follows this convention for expressing linear models. 
+
+_________________
+
+## Sebert & Montgomery & Rollier (1998) Algorithm
+Sebert & Montgometry & Rollier (smr98) algorithm starts with an ordinary least squares estimation for a given model and data. Residuals and fitted responses are calculated using the estimated model. A hierarchical clustering analysis is applied using standardized residuals and standardized fitted responses. The tree structure of clusters are cut using a threshold, e.g Majona criterion, as suggested by the authors. It is expected that the subtrees with relatively small number of observations are declared to be clusters of outliers.
+
+Hawkings & Bradu & Kass dataset has 4 variables and 75 observations. The observation 1-14 are known to be outliers. In the example below, we create an regression setting using the formula ```y ~ x1 + x2 + x3``` and ```hbk``` dataset. ```smr98``` is directly applied on this setting.  
+
 ```julia
 julia> using LinRegOutliers
 julia> # Regression setting for Hawkins & Bradu & Kass data
 julia> reg = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk)
 julia> smr98(reg)
-14-element Array{Int64,1}:
-  1
-  2
-  3
-  4
-  5
-  6
-  7
-  8
-  9
- 10
- 11
- 12
- 13
- 14
+Dict{String,Array{Int64,1}} with 1 entry:
+  "outliers" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
 ```
 
+The Julia method ```smr98()``` returns a ```Dict``` object with produced output. In this case, the single output is indices of detected outlying observations. 
+The algorithm successfully detects the outliers.
+
+________________________
+## Peña and Yohai (1995) 
+
+Peña and Yohai (```py1995```) algorithm starts by constructing an influence matrix using results of an ordinary least squares estimate for a given model and data. In the second stage, the eigen structure of the influence matrix is examined for detecting suspected subsets of data. 
+
+Here is an example of ```py95``` method applied on the ```hbk``` data. The method returns a ```Dict``` object with keys ```outliers``` and ```suspected.sets```. An usual researcher may directly focused on the ```outliers``` indices. The method reports the observations 1-14 are outliers.
 
 ```julia
 julia> py95(reg)["outliers"]
-14-element Array{Int64,1}:
-  1
-  2
-  3
-  4
-  5
-  6
-  7
-  8
-  9
- 10
- 11
- 12
- 13
- 14
-
+Dict{Any,Any} with 2 entries:
+  "outliers"       => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
[…]
