### 2020-12-31 0588a2f
add image for lts and phones data to examples.md
