### 2020-12-31 393c203
update detectOutliers multi-method
