### 2020-12-31 jbytecode/LinRegOutliers#41 @jbytecode
Dear reviewer @salleuska,

[example page](https://github.com/jbytecode/LinRegOutliers/blob/master/examples.md) is updated with additional information about selected methods. An introduction section is added for better explanation of basic calling convention. The other calling convention is also mentioned. The datasets used in examples are also clarified.
