### 2021-01-01 22c49c1
update documentation for atkinson94
