### 2021-01-01 4a1078a
update documentation for hadi1992 and hadi1994
