### 2021-01-02 6ba85cb
update documentation for ks89
