+++ src/ks89.jl
@@ -1,12 +1,15 @@
 """
     ks89RecursiveResidual(setting; indices, k)
 Calculate recursive residual for the given regression setting and observation.
+
 # Arguments
 - `setting::RegressionSetting`: RegressionSetting object with a formula and dataset.
 - `indices::ArrayInt,1`: Indices of observations used in the linear model.
 - `k::Int`: Observation indice the recursive residual is calculated for.
+
 # Notes
     This is a helper function for the ks89 function and it is not directly used.
+
 # References
 Kianifard, Farid, and William H. Swallow. "Using recursive residuals, calculated on
 adaptively-ordered observations, to identify outliers in linear regression."
@@ -30,10 +33,22 @@ end
 
 """
     ks89(setting; alpha = 0.05)
+
 Perform the Kianifard & Swallow (1989) algorithm for the given regression setting.
+
+
 # Arguments
 - `setting::RegressionSetting`: RegressionSetting object with a formula and dataset.
 - `alpha::Float64`: Optional argument of the probability of rejecting the null hypothesis.
+
+# Description
+The algorithm starts with a clean subset of observations. This initial set is then enlarged 
+using recursive residuals. When the calculated statistics exceeds a threshold it terminates. 
+
+
+# Output
+- `["outliers]`: Array of indices of outliers.
+
 # Examples
 ```julia-repl
 julia> reg0001 = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
