+++ src/py95.jl
@@ -148,6 +148,16 @@ Perform the Pena & Yohai (1995) algorithm for the given regression setting.
 # Arguments
 - `setting::RegressionSetting`: RegressionSetting object with a formula and dataset.
 
+# Description 
+The algorithm starts by constructing an influence matrix using results 
+of an ordinary least squares estimate for a given model and data. In the second stage, 
+the eigen structure of the influence matrix is examined for detecting suspected subsets of data.
+
+# Output 
+- `["outliers"]`: Array of indices of outliers
+- `["suspected.sets"]`: Arrays of indices of observations for corresponding eigen value of the influence matrix.
+
+
 # Examples
 ```julia-repl
 julia> reg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
