### 2021-01-02 3fa9132
update documentation for satman2015
