### 2021-01-02 c09d17c
update documentation for smr98
