### 2021-01-02 jbytecode/LinRegOutliers#40 @jbytecode
Dear reviewer @salleuska,

Basically, each algorithm had a documentation sections with fields `Arguments`, `Examples`, and `References`.

Now, I added two new sections `Description` and `Output`. In `Description` section, the corresponding algorithm is described with a few sentences. In `Output` section, keys of the returned `Dict` object are listed and explained.

Reference manual is updated.

All of the methods return a Dict object even the Dict has only one key.
