### 2021-01-02 2ed3a9b
update paper and bibtex
