+++ paper/paper.bib
@@ -414,4 +414,11 @@ and Schader, Martin},
   doi       = {10.1080/03610918.2011.598989}
 }
 
-
+@report{mahalanobis,
+  author    = {Mahalanobis, P. C.},
+  year      = {1930},
+  title     = {On tests and measures of group divergence. Part 1: Theoretical formulae},
+  publisher = {Journal and Proceedings Asiatic Society of Bengal (New Series)},
+  volume    = {26},
+  pages     = {541–580}
+}
