### 2021-01-02 ebaf556
update paper and bibtex
