### 2021-01-02 2dab08d
update paper and bibtex
