### 2021-01-02 131d750
update paper and bibtex
