### 2021-01-02 6fcc7c1
update paper and bibtex
