+++ paper/paper.bib
@@ -418,7 +418,7 @@ and Schader, Martin},
   author    = {Mahalanobis, P. C.},
   year      = {1930},
   title     = {On tests and measures of group divergence. Part 1: Theoretical formulae},
-  publisher = {Journal and Proceedings Asiatic Society of Bengal (New Series)},
+  publisher = {Journal & Proceedings Asiatic Society of Bengal (New Series)},
   volume    = {26},
   pages     = {541–580}
 }
