### 2021-01-02 #2892 @rMassimiliano
I really like the new example section. I have some comments on the main paper and on the example section in the repository in the following.  I tested the code on Julia 1.5.3 on Ubuntu.

**Paper**
Following line-by-line the code in the section **Installation and basic usage**, the examples at line 70 and 71 return `Array` not `Dict`. Make sure the results shown in the paper are consistent with the code in the repository.

**Repository** (`examples.md`)

**Sebert & Montgomery & Rollier (1998) Algorithm:**
```
smr98(reg); ## return `Array` not `Dict`
```
**Peña and Yohai (1995)**
I would change  'suspected subsets' into 'subset of potential outliers'.
in the code snippets I think `py95(reg)["outliers"]` should be `py95(reg)`

```
py95(reg)
Dict{Any,Any} with 2 entries:
  "outliers"       => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "suspected.sets" => Set([[14, 13], [43, 54, 24, 38, 22], Int64[], [58, 66, 32,…
```

While the current version would return

```
py95(reg)["outliers"]
14-element Array{Int64,1}:
  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
```
**Least Trimmed Squares Regression**
I would add the code to reproduce the plot in a snippet (or a link to a separate file if too long)
