### 2021-01-02 5425f4d
update examples.md
