### 2021-01-02 #2892 @jbytecode
Dear reviewer @rMassimiliano,

Thank you very much for your comments.

- In the paper, lines 70-71, the method `smr98` (and all of the methods exported from the package) now return a `Dict` after pushing many commits. Could you please re-check? Or please tell me if I am unlucky missing something.
- Peña and Yohai (1995) section is corrected.
- A code snipped is added for drawing LTS line and Phone data.
