### 2021-01-03 #2892 @rMassimiliano
Dear @jbytecode,
the functions on your repository  ` https://github.com/jbytecode/LinRegOutliers`  are indeed updated and return `Dict` as expected.

However, the new version has not been added to Julia package registry `JuliaRegistries/General`. This is the version installed when using `Pkg.add()`, as suggested in the **Installation** section of the repo.

I would recommend to add the new version to `JuliaRegistries/General', and also to change the package version, given that the return type of some functions is changed.
