### 2021-01-03 #2892 @jbytecode
Dear reviewer @rMassimiliano,

New version is ready in Julia package registry now.

Thank you.
