### 2021-01-04 #2892 @salleuska
> Dear reviewer @salleuska,
>
> * Methods are consistently returning a `Dict` object even the result includes only one item.
> * All exported methods have now `Description` and `Output` sections in their documentation. In the former, we give a short and brief introduction, how it runs, what does it do, etc., without getting deep into the details. In the latter, we present what the method returns and what is the meaning of these items.
> * Examples.md is updated.
> * Documentation is updated.
> * Summary is fixed.
> * Statement of field section is changed a little bit. There are now 5 categories, including the visual methods, and listed in separate tables, each for diagnostics, direct methods, robust methods, methods for multivariate data, and visual methods.
>
> I think, these changes cover all your suggestions. Please, let me know, if there is anything missing or wrong.
>
> Thank you.

Great, happy to see the changes! I don't have any other suggestion at this point.

@mikldk my review is now complete, and I would recommend acceptance.
