### 2021-01-04 jbytecode/LinRegOutliers#40 @jbytecode
Dear reviewer @salleuska

 can we now close the issues if they are okay?

Thanks.
