### 2021-01-04 #2892 @rMassimiliano
> Dear reviewer @rMassimiliano,
>
> New version is ready in Julia package registry now.
>
> Thank you.

Great! Everything works as expected now.

@mikldk I have no further comments, and I would also recommend acceptance.
