### 2021-01-04 #2892 @mikldk
> > @rMassimiliano, please carefully go though the checklist at the top, too (and check items off as you go).
>
> Dear @mikldk I went through the checklist, but I am unable to check the items. I probably forgot to accept the invite `https://github.com/openjournals/joss-reviews/invitations` which is now expired. Can you help me with this?

@rMassimiliano I have tried removing and adding you. Can you please try again?
