### 2021-01-04 #2892 @mikldk
@jbytecode :

`README` (in the repo):

* Typo "mutliple regression"
* Headline "Want to have contributions?" -- not Title Case like the other headings in the `README`, and the wording is a bit strange as you as the maintainer want to have contributions, and potential contributors would want to make contributions, I guesss. Consider changing the heading to just e.g. "Contributions".
* A full stop is missing at the first item in "Contact & Communication"

Paper:

* L14-15: The formulation "Our package covers a significant portion of the literature" (maybe also found elsewhere in the package's documentation/paper?) is imprecise; try to reformulate.
* L21-22: "We then seek to find the linear model"; consider replacing **the** linear model with **a** linear model (I am thinking of what objective function that is minimised)

Finally:

* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
