### 2021-01-04 #2892 @jbytecode
Dear editor @mikldk,

The latest update includes these items:

- `README.md` is corrected
- L14-15: It is replaced by the new sentence "Our package implements many algorithms and diagnostics for model fitting with outliers under a single interface, which allows users to quickly try many different methods with reasonable default settings, while also providing a good starting framework for researchers who may want to extend the package with novel methods."
- L21-22: Corrected
- A table name is changed from Algorithm(s) to Algorithm
- We double checked for typos

Finally, here is the zenodo essentials:
- DOI: `10.5281/zenodo.4415671`
- Bagdet: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4415671.svg)](https://doi.org/10.5281/zenodo.4415671)
- Here is the [release](https://github.com/jbytecode/LinRegOutliers/releases/tag/0.8.4) with version id `0.8.4`

Thank you.
