### 2021-01-05 #2892 @mikldk
@jbytecode There are some confusing about the tags/releases in your [repository](https://github.com/jbytecode/LinRegOutliers/releases) (`v0.8.4` vs `0.8.4`, see e.g. [the diff](https://github.com/jbytecode/LinRegOutliers/compare/v0.8.4...0.8.4)).

If changed have been made since 0.8.4, then I recommend making this JOSS submission version 0.8.5 as the `v` prefix is not a part of the semantic versioning.
