### 2021-01-05 #2892 @jbytecode
It is just because the v0.8.4 was automatically created by the tag bot after publishing the latest revision in Julia repos. So I set the name of release 0.8.4, without a v prefix.

Now, I am trying to create a new release, but zenodo throws an error of "New version's files must differ from all previous versions."

what to do now ?
